Implement handover, improve a lot of things

+++ src/main/java/org/saltyrtc/client/signaling/Initiator.java
@@ -24,7 +24,7 @@ public class Initiator extends Peer {
         this.handshakeState = InitiatorHandshakeState.NEW;
     }
 
-    public int getId() {
+    public short getId() {
         return 0x01;
     }
